cleanDetection/requirements.txt
absl-py==2.1.0
altgraph==0.17.4
auto-py-to-exe==2.43.3
blinker==1.7.0
boto3==1.34.83
botocore==1.34.83
bottle==0.12.25
bottle-websocket==0.2.9
certifi==2023.7.22
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cvzone==1.6.1
cx_Freeze==6.15.16
cx_Logging==3.2.0
cycler==0.10.0
Cython==3.0.10
detect==2020.12.3
Eel==0.16.0
et-xmlfile==1.1.0
expecttest==0.2.1
filelock==3.13.4
fire==0.6.0
Flask==3.0.3
fonttools==4.51.0
fsspec==2024.3.1
future==1.0.0
gevent==24.2.1
gevent-websocket==0.10.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
grpcio==1.62.1
huggingface-hub==0.22.2
idna==2.10
itsdangerous==2.2.0
Jinja2==3.1.3
jmespath==1.0.1
kiwisolver==1.4.5
lapx==0.5.7
lief==0.14.1
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.8.4
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
onnx==1.16.0
onnxscript==0.1.0.dev20240418
opencv-python==4.7.0.72
opencv-python-headless==4.8.0.74
openpyxl==3.1.2
packaging==24.0
pandas==2.2.2
pefile==2023.2.7
pillow==10.3.0
protobuf==5.26.1
psutil==5.9.8
py-cpuinfo==9.0.0
pybboxes==0.1.6
pycparser==2.22
pyinstaller==6.6.0
pyinstaller-hooks-contrib==2024.4
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-magic==0.4.27
pytz==2024.1
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0.1
requests==2.31.0
requests-toolbelt==1.0.0
roboflow==1.1.27
s3transfer==0.10.1
sahi==0.11.15
scipy==1.13.0
seaborn==0.13.2
setuptools==69.2.0
shapely==2.0.3
six==1.16.0
smmap==5.0.1
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
termcolor==2.4.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.2
typing_extensions==4.11.0
tzdata==2024.1
ultralytics==8.1.47
urllib3==2.2.1
Werkzeug==3.0.2
wheel==0.43.0
whichcraft==0.6.1
yolov5==7.0.13
yolov8==0.0.2
zope.event==5.0
zope.interface==6.3
